move codes to utils
